simplify dd_init.php <> functions.php files
